requirements.txt
python-dotenv>=0.19.0
python-telegram-bot>=13.7
aiohttp>=3.8.0
beautifulsoup4>=4.10.0
undetected-chromedriver>=3.0.0
fake-useragent>=0.1.11
yarl>=1.7.0
watchdog>=2.1.0 